Compute avatar color of users and apply foreground color to the sender displayname.
